% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationData.R
\docType{data}
\name{classificationData}
\alias{classificationData}
\title{Two groups of Gaussian processes with different mean values}
\format{
A matrix with n = 100 functions by columns and t = 200 evaluation points by row.
The first 50 are G1 and second 50 curves are G2 that differs in the mean value.
}
\usage{
classificationData
}
\description{
Two groups of Gaussian processes with different mean values
}
\examples{
matplot(classificationData, type = "l")
}
\references{
Elías, Antonio, Jiménez, Raúl and Yukich, Joe (2020). Localization processes for functional data analysis (submitted).
}
\keyword{datasets}
