% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Functional Gaussian processes.}
\format{
A matrix with n = 1000 functions by columns and t = 100 evaluation points by row.
}
\usage{
exampleData
}
\description{
Functional Gaussian processes.
}
\examples{
matplot(exampleData, type = "l")
}
\references{
Elías, Antonio, Jiménez, Raúl and Yukich, Joe (2020). Localization processes for functional data analysis (submitted).
}
\keyword{datasets}
