% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizationDistances.R
\name{localizationDistances}
\alias{localizationDistances}
\title{Localization distances}
\usage{
localizationDistances(y, y0)
}
\arguments{
\item{y}{matrix p by n, being n the number of functions and p the number of grid points.}

\item{y0}{focal curve (index or character name).}
}
\value{
a vector of length (n-1), being the localization distance of its corresponding order.
}
\description{
Compute the localization distances of order k of the curve \code{y0}.
}
\examples{
localizationDistances_1 <- localizationDistances(exampleData, y0 = "1")

}
\references{
Elías, Antonio, Jiménez, Raúl and Yukich, Joe (2020). Localization processes for functional data analysis (submitted).
}
