% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizationProcesses.R
\name{localizationProcesses}
\alias{localizationProcesses}
\title{Localization processes}
\usage{
localizationProcesses(y, y0)
}
\arguments{
\item{y}{matrix p by n, being n the number of functions and p the number of grid points.}

\item{y0}{focal curve index or name}
}
\value{
a list with one element, \code{lc}, a matrix of size p x (n-1), being the (n-1) columns the localization processes of its corresponding order.
}
\description{
Compute the localization processes of order k of the curve \code{y0}.
}
\examples{
localizationProcesses_1 <- localizationProcesses(exampleData, y0 = "1")

}
\references{
Elías, Antonio, Jiménez, Raúl and Yukich, Joe (2020). Localization processes for functional data analysis (submitted).
}
