% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localizationStatistics.R
\name{localizationStatistics}
\alias{localizationStatistics}
\title{Localization Distances Statistics}
\usage{
localizationStatistics(y, robustify = TRUE, whisker_rule)
}
\arguments{
\item{y}{matrix p by n, being n the number of functions and p the number of grid points.}

\item{robustify}{if TRUE the mean and standard deviation are estimated with a the trimmed sample. Default is TRUE.}

\item{whisker_rule}{Range parameter for the univariate boxplot detection rule. Default = 3.}
}
\value{
a list with the localization distances of each function (localizationDistances),
the estimated mean (mean) and standard deviation (sd).
}
\description{
Estimate the mean and standard deviation of the localization distances mean.
}
\examples{
data(exampleData)
localizationDistances_1 <- localizationDistances(X, y0 = "1")

}
\references{
Elías, Antonio, Jiménez, Raúl and Yukich, Joe (2020). Localization processes for functional data analysis (submitted).
}
