% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierData.R
\docType{data}
\name{outlierData}
\alias{outlierData}
\title{Functional Gaussian processes with outliers.}
\format{
A matrix with n = 54 functions by columns and t = 200 evaluation points by row.
The last 4 observations are two shape and two magnitude outliers.
}
\usage{
outlierData
}
\description{
Functional Gaussian processes with outliers.
}
\examples{
matplot(outlierData, type = "l")
}
\references{
Elías, Antonio, Jiménez, Raúl and Yukich, Joe (2020). Localization processes for functional data analysis (submitted).
}
\keyword{datasets}
