% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlyingLocalizationDistances.R
\name{outlierLocalizationDistance}
\alias{outlierLocalizationDistance}
\title{Outlying localization distances}
\usage{
outlierLocalizationDistance(X, local_rule = 0.9, whisker_rule = 3)
}
\arguments{
\item{X}{matrix p by n, being n the number of functions and p the number of grid points.}

\item{local_rule}{Local distance rule: the method marks a curve as outlier if
its k order localization distances are outliers in more than local_rulex100% of the k-order univariate boxplots.
Default is 0.90 so a function must be at least an outlier in 90% of the k-order localization distances.}

\item{whisker_rule}{Parameter for the whiskers of the univariate boxplot of the
localization distances of order kth. Default value is 3.}
}
\value{

}
\description{
Compute the localization distances of order k of the curve \code{y0}.
}
\examples{
data(outliersData)
outliers <- outlierLocalizationDistance(X, local_rule = 0.9, whisker_rule = 3)
outliers$outliers_ld_rule
}
\references{
Elías, Antonio, Jiménez, Raúl and Yukich, Joe (2020). Localization processes for functional data analysis (submitted).
}
